requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
cryptography==2.3.1
ephem==3.7.6.0
future==0.16.0
idna==2.7
pycparser==2.19
PySocks==1.6.8
python-telegram-bot==11.1.0
PyYAML==3.13
six==1.11.0
urllib3==1.23
